refactor(benchpress): delete perf_util.ts and moved it's content to e2e_util.ts
* This change makes the external api more closely resemble the internal one so that syncing benchpress drivers can be as simple as possible.
