+++ modules/benchmarks/src/largeform/largeform.perf-spec.ts
@@ -8,8 +8,7 @@
 
 import {$} from 'protractor';
 
-import {verifyNoBrowserErrors} from '../../../e2e_util/e2e_util';
-import {runBenchmark} from '../../../e2e_util/perf_util';
+import {runBenchmark, verifyNoBrowserErrors} from '../../../e2e_util/e2e_util';
 
 interface Worker {
   id: string;
