+++ modules/benchmarks/src/largetable/largetable.perf-spec.ts
@@ -7,7 +7,7 @@
  */
 
 import {$} from 'protractor';
-import {runBenchmark, verifyNoBrowserErrors} from '../../../e2e_util/perf_util';
+import {runBenchmark, verifyNoBrowserErrors} from '../../../e2e_util/e2e_util';
 
 interface Worker {
   id: string;
