+++ modules/benchmarks/src/styling/BUILD.bazel
@@ -5,7 +5,7 @@ load("//tools:defaults.bzl", "ts_library")
 ts_library(
     name = "tests_lib",
     testonly = True,
-    srcs = ["styling_perf.spec.ts"],
+    srcs = ["styling.perf-spec.ts"],
     tsconfig = "//modules/benchmarks:tsconfig-e2e.json",
     deps = [
         "//modules/e2e_util",
