+++ modules/benchmarks/src/styling/styling.perf-spec.ts
@@ -7,8 +7,9 @@
  */
 
 import {$, by, element} from 'protractor';
-import {openBrowser, verifyNoBrowserErrors} from '../../../e2e_util/e2e_util';
-import {runBenchmark} from '../../../e2e_util/perf_util';
+
+import {openBrowser, runBenchmark, verifyNoBrowserErrors} from '../../../e2e_util/e2e_util';
+
 
 /** List of possible scenarios that should be tested.  */
 const SCENARIOS = [
